Isolate content-sensitive e2e specs from cross-spec thread reuse
All specs authenticate as the same OIDC test-user and the backend
scopes thread search to that identity, so getOrCreateThread()'s
latest-idle reuse lets one spec inherit another's thread — including
persisted `phase` state that reroutes the ai-mock's offset-matched
responses (e.g. "PHASE_VISIBLE: draft" instead of the echo the
Edit/Regenerate/Cancellation assertions depend on).

Hoist phase.spec.ts's gotoFreshThread() into fixtures/backend.ts and
use it in chat.spec.ts's content-sensitive describe blocks. Specs that
deliberately exercise the getOrCreateThread redirect flow are left
untouched. Also guard the Regenerate test's captured Bearer token with
an actionable expect instead of a bare non-null assertion.

Addresses review feedback from Codex (cross-spec thread isolation)
and Copilot (capturedToken guard) on PR #259.
